requirements.txt
crewai==0.1.24
unstructured==0.10.25
pyowm==3.3.0
tools>=0.1.9
python-dotenv==1.0.0
langchain-google-vertexai>=1.0.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
gunicorn>=21.2.0
